merlin: remove data Python backend
