+++ src/libertem_live/detectors/merlin/__init__.py
@@ -1,9 +1,8 @@
-from .data import MerlinDataSource
 from .control import MerlinControl
 from .acquisition import MerlinAcquisition
 from .connection import MerlinConnectionBuilder, MerlinDetectorConnection
 
 __all__ = [
-    "MerlinDataSource", "MerlinControl", "MerlinAcquisition",
+    "MerlinControl", "MerlinAcquisition",
     "MerlinConnectionBuilder", "MerlinDetectorConnection",
 ]
